world convert progress info
